% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tables.R
\name{d4p_create_tables}
\alias{d4p_create_tables}
\title{Create databasing4patches tables}
\usage{
d4p_create_tables(con, tables = "all")
}
\arguments{
\item{con}{A DBI database connection}

\item{tables}{Which tables to create ("all" is default)}
}
\description{
Create databasing4patches tables
}
\examples{
con <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")
d4p_create_tables(con)
DBI::dbDisconnect(con)
}
