% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{DB4Patches}
\alias{DB4Patches}
\title{DB4Patches Object}
\value{
Nothing
}
\description{
Interface for a hiking tracks, bagged points, and collected trails database.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{con}}{DB4Patches. Read-only.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-db4patches_object-new}{\code{DB4Patches$new()}}
\item \href{#method-db4patches_object-establish_connection}{\code{DB4Patches$establish_connection()}}
\item \href{#method-db4patches_object-disconnect}{\code{DB4Patches$disconnect()}}
\item \href{#method-db4patches_object-finalize}{\code{DB4Patches$finalize()}}
\item \href{#method-db4patches_object-create_tables}{\code{DB4Patches$create_tables()}}
\item \href{#method-db4patches_object-tbl}{\code{DB4Patches$tbl()}}
\item \href{#method-db4patches_object-track_exists}{\code{DB4Patches$track_exists()}}
\item \href{#method-db4patches_object-add_track}{\code{DB4Patches$add_track()}}
\item \href{#method-db4patches_object-point_exists}{\code{DB4Patches$point_exists()}}
\item \href{#method-db4patches_object-add_point}{\code{DB4Patches$add_point()}}
\item \href{#method-db4patches_object-add_point_sfdf}{\code{DB4Patches$add_point_sfdf()}}
\item \href{#method-db4patches_object-add_point_set}{\code{DB4Patches$add_point_set()}}
\item \href{#method-db4patches_object-clone}{\code{DB4Patches$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-new"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-new}{}}}
\subsection{Method \code{new()}}{
Create new DB4Patches object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$new(dbargs_list, create_tables = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbargs_list}}{A list of parameters to create the database}

\item{\code{create_tables}}{Create tables in the database (default: FALSE)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-establish_connection"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-establish_connection}{}}}
\subsection{Method \code{establish_connection()}}{
Create DB4Patches object and connect to a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$establish_connection(dbargs_list)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dbargs_list}}{A list of parameters to create the database}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
Disconnect from a database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-finalize}{}}}
\subsection{Method \code{finalize()}}{
Clean up after removal of DB4Patches object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-create_tables"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-create_tables}{}}}
\subsection{Method \code{create_tables()}}{
Generate a data frame interface to a database table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$create_tables()}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl_name}}{A database table name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-tbl"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-tbl}{}}}
\subsection{Method \code{tbl()}}{
Generate a data frame interface to a database table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$tbl(tbl_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl_name}}{A database table name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-track_exists"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-track_exists}{}}}
\subsection{Method \code{track_exists()}}{
Check if a track is already in the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$track_exists(track_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{track_id}}{The track hash}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-add_track"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-add_track}{}}}
\subsection{Method \code{add_track()}}{
Add a track to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$add_track(track)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{track}}{A GPX-style track data frame}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-point_exists"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-point_exists}{}}}
\subsection{Method \code{point_exists()}}{
Check if a point is already in the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$point_exists(lat, lon, plusminus = 0.01, conn = self$con)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{The point latitude}

\item{\code{lon}}{The point longitude}

\item{\code{plusminus}}{The plus/minus degree range difference allowed}

\item{\code{conn}}{(Internal) Database pool connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-add_point"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-add_point}{}}}
\subsection{Method \code{add_point()}}{
Add a point to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$add_point(
  lat,
  lon,
  name = NA,
  elev_m = NA,
  elev_ft = NA,
  conn = self$con
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lat}}{Latitude in degrees in WGS84 CRS}

\item{\code{lon}}{Longitude in degrees in WGS84 CRS}

\item{\code{name}}{point name}

\item{\code{elev_m}}{Elevation in meters}

\item{\code{elev_ft}}{Elevation in feet}

\item{\code{conn}}{(Internal) Database pool connection}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-add_point_sfdf"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-add_point_sfdf}{}}}
\subsection{Method \code{add_point_sfdf()}}{
Add point(s) in sf data frame to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$add_point_sfdf(
  sfdf,
  name_col = NA,
  elev_m_col = NA,
  elev_ft_col = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{sfdf}}{sf data frame}

\item{\code{name_col}}{Name of name column}

\item{\code{elev_m_col}}{Name of meters elevation column}

\item{\code{elev_ft_col}}{Name of feet elevation column}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-add_point_set"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-add_point_set}{}}}
\subsection{Method \code{add_point_set()}}{
Add a point set to the database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$add_point_set(
  name,
  points_type,
  version,
  single = FALSE,
  single_type = NA,
  depreciated = NA,
  notes = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{Point set name}

\item{\code{points_type}}{"peaks" | "fire towers"}

\item{\code{version}}{Start date for this version of the point set}

\item{\code{single}}{To be completed within a single time frame}

\item{\code{single_type}}{Time frame of month, season, or year}

\item{\code{depreciated}}{Date (if) this version has been superseded/ended}

\item{\code{notes}}{Notes}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-db4patches_object-clone"></a>}}
\if{latex}{\out{\hypertarget{method-db4patches_object-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB4Patches$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
