% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_bagging.R
\name{make_peak_bagger_function}
\alias{make_peak_bagger_function}
\title{Make peak bagger function for given list of peaks}
\usage{
make_peak_bagger_function(peaks_sf)
}
\arguments{
\item{peaks_sf}{An sf data frame of peaks}
}
\value{
A function that takes a GPX track as an argument
}
\description{
Given an sf data frame with a list of peaks, this will create a function
that determines which peaks a given track intersects. The sf data frame
must have a "peak_id" column and the required sf "geometry" column and
should be in the GPS (4326) coordinate reference system.
}
\examples{
peak_bagger <- make_peak_bagger_function(peaks)
peak_bagger(kaaterskill$tracks[[1]])
}
