% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interface.R
\name{make_pool_warning_go_away}
\alias{make_pool_warning_go_away}
\title{Nonsense packaging issue function}
\usage{
make_pool_warning_go_away(con)
}
\arguments{
\item{con}{A database connection}
}
\value{
An API version.
}
\description{
The [pool] package is used, but indirectly. This function exists
only to kill an annoying packaging `check` warning.
}
